+++ src/compiler/directive.js
@@ -17,11 +17,11 @@ export default {
         const directive = hyphensToCamelCase( node.name ) + 'Directive' +
             figure.uniqid( 'directive_name' );
 
-        figure.addDirective( sourceNode( node.loc, `    let ${directive};` ) );
+        figure.addDirective( sourceNode( node.loc, `    let ${directive};` ), node.name );
         figure.addRenderActions(
             sourceNode( node.loc, [
-                `        if ( ${directive} === undefined ) {\n`,
-                `            ${directive} = new this.directives.${node.name}( ${figure.getPathToDocument()} );\n`,
+                `        if ( !${directive} ) {\n`,
+                `            ${directive} = new ${figure.getDirectiveAlias( node.name )}( ${figure.getPathToDocument()} );\n`,
                 '        }\n',
                 `        ${directive}.bind( ${parent.reference} );`
             ] )