+++ src/transform/useblock.js
@@ -1,15 +1,29 @@
+import { getStringLiteralValue } from '../utils';
 import { visit } from '../visitor';
 import { HTMLElements, SVGElements } from '../compiler/element';
 import { ast } from '../parser';
+
+const uniqcounters = {};
 
 export function useBlockWithCompound( ast ) {
     visit( ast, {
         UseBlockStatement: ( node ) => {
             if ( !node.withCustom ) {
+                addUniqBlockDataName( node );
                 handleCustomComponentsInBlock( node );
             }
         }
     } );
+}
+
+function addUniqBlockDataName( node ) {
+    const blockName = getStringLiteralValue( node.name  );
+    if ( !uniqcounters[ blockName ] ) {
+        uniqcounters[ blockName ] = 0;
+    }
+    node.identifier = new ast.IdentifierNode(
+        `dataForBlock_${blockName}_${uniqcounters[ blockName ]++}`
+    );
 }
 
 
@@ -20,16 +34,16 @@ function handleCustomComponentsInBlock( blockNode ) {
                 !HTMLElements.includes( node.name ) &&
                 !SVGElements.includes( node.name )
             ) {
-                addBlockDataSpreadAttribute( node );
+                addBlockDataSpreadAttribute( blockNode, node );
             }
         }
     } );
 }
 
-function addBlockDataSpreadAttribute( node ) {
+function addBlockDataSpreadAttribute( blockNode, node ) {
     node.attributes = node.attributes.concat(
         new ast.SpreadAttributeNode( [
-            new ast.IdentifierNode( 'blockData' )
+            new ast.IdentifierNode( blockNode.identifier.name )
         ] )
     );
 }